# docs/swagger.yaml
basePath: /
definitions:
  controller.LoginReply:
    properties:
      expired_at:
        description: token失效时间
        type: integer
      id:
        description: 用户id
        type: integer
      token:
        description: 颁发的token
        type: string
    type: object
  models.UserListReply:
    properties:
      add_time:
        description: 注册时间
        type: integer
      id:
        description: 用户id
        type: integer
      mobile:
        description: 用户手机号
        type: string
      name:
        description: 用户名
        type: string
      status:
        description: 用户状态
        type: integer
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: qiujun@sina.com
    url: http://www.swagger.io/support
  description: go web framework demo
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: web_app
  version: "1.0"
paths:
  /user/list:
    get:
      consumes:
      - application/json
      description: 可分页查询用户信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: page_num
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserListReply'
            type: array
      security:
      - ApiKeyAuth: []
      summary: 用户信息列表接口
      tags:
      - 用户相关接口
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户通过手机号和密码登录
      parameters:
      - in: query
        name: mobile
        required: true
        type: string
      - in: query
        maxLength: 20
        minLength: 6
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.LoginReply'
      summary: 用户登录接口
      tags:
      - 用户相关接口
  /users/register:
    post:
      consumes:
      - application/json
      description: 用户填写用户名，手机号，密码并确认密码无误即可注册
      parameters:
      - description: 手机号，必填
        in: query
        name: mobile
        required: true
        type: string
      - description: 密码，长度范围6到20位
        in: query
        maxLength: 20
        minLength: 6
        name: password
        required: true
        type: string
      - description: 确认密码，必须与密码一致
        in: query
        maxLength: 20
        minLength: 6
        name: re_password
        required: true
        type: string
      - description: 用户名，必填
        in: query
        name: user_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户注册接口
      tags:
      - 用户相关接口
swagger: "2.0"